renamed exercises to labs
