+++ material/src/test/scala/fhj/swengb/labs/lab02/Gruppe1Test.scala
@@ -1,4 +1,4 @@
-package fhj.swengb.exercises.ex02
+package fhj.swengb.labs.lab02
 
 import org.junit.Assert._
 import org.junit.Test
